[r2r][p 222] react & maintenance; project structure, split code into directories and files
