+++ src/App/Button/index.tsx
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+export const StyledButton = styled.button`
+  background: transparent;
+  border: 2px solid #222;
+  border-radius: 4px;
+  padding: 4px;
+  cursor: pointer;
+
+  :hover {
+    color: #eee;
+    background: #222;
+    border: 4px solid #888;
+  }
+`;
+
