+++ src/App/Data/index.ts
@@ -1,0 +1,14 @@
+const hnBaseApi = 'https://hn.algolia.com/api/v1';
+
+const noItem = '';
+
+type Story = {
+    objectID: number;
+    url: string;
+    title: string;
+    author: string;
+    num_comments: number;
+    points: number;
+};
+
+export { hnBaseApi, noItem, Story };
